requirements.txt
streamlit
faiss-cpu
sentence-transformers
transformers
PyPDF2
opencv-python
pytesseract
